refactor(frontend): separate page entry renderers
